@@ -5,7 +5,7 @@ An object is plain if it's created by either `{}`, `new Object()`, or `Object.cr
 
 @example
 ```
-import isPlainObject = require('is-plain-obj');
+import isPlainObject from 'is-plain-obj';
 
 isPlainObject({foo: 'bar'});
 //=> true
@@ -24,6 +24,4 @@ isPlainObject(new Unicorn());
 //=> false
 ```
 */
-declare function isPlainObject<Value = unknown>(value: unknown): value is Record<string | number | symbol, Value>;
-
-export = isPlainObject;
+export default function isPlainObject<Value>(value: unknown): value is Record<string | number | symbol, Value>;
