@@ -1,10 +1,8 @@
-'use strict';
-
-module.exports = value => {
+export default function isPlainObject(value) {
 	if (Object.prototype.toString.call(value) !== '[object Object]') {
 		return false;
 	}
 
 	const prototype = Object.getPrototypeOf(value);
 	return prototype === null || prototype === Object.prototype;
-};
+}
